Move annotation styles back to component folder
