+++ projects/js-packages/charts/src/components/line-chart/index.tsx
@@ -5,3 +5,4 @@ export type {
 	TooltipDatum,
 	CurveType,
 } from './line-chart';
+export type { AnnotationStyles } from './types';
