+++ projects/js-packages/charts/src/types.ts
@@ -1,7 +1,4 @@
-import { CircleSubjectProps } from '@visx/annotation/lib/components/CircleSubject';
-import { ConnectorProps } from '@visx/annotation/lib/components/Connector';
-import { LabelProps } from '@visx/annotation/lib/components/Label';
-import { LineSubjectProps } from '@visx/annotation/lib/components/LineSubject';
+import type { AnnotationStyles } from './components/line-chart';
 import type { AxisScale, Orientation, TickFormatter, AxisRendererProps } from '@visx/axis';
 import type { LegendShape } from '@visx/legend/lib/types';
 import type { ScaleInput, ScaleType } from '@visx/scale';
@@ -98,16 +95,6 @@ export type BaseLegendItem = {
 	group?: string;
 	index?: number;
 	overrideColor?: string;
-};
-
-export type AnnotationStyles = {
-	circleSubject?: Omit< CircleSubjectProps, 'x' | 'y' > & { fill?: string };
-	lineSubject?: Omit< LineSubjectProps, 'x' | 'y' >;
-	connector?: Omit< ConnectorProps, 'x' | 'y' | 'dx' | 'dy' >;
-	label?: Omit< LabelProps, 'title' | 'subtitle' | 'x' | 'y' > & {
-		x?: number | 'start' | 'end';
-		y?: number | 'start' | 'end';
-	};
 };
 
 /**
